+++ runtime/gc/umarray-allocate.c
@@ -241,7 +241,7 @@ pointer GC_arrayAllocate(GC_state s,
 	if (DEBUG_MEM) {
         fprintf(stderr, "%d] GC_arrayAllocate: numElements: %zd (%zd b/e), "
 						"numElsPerChunk: %zd, numLeaves: %zd, numChunks: %zd, treeHeight: %zd, tcn: %f\n",
-                PTHREAD_NUM, numElements, bytesPerElement, numElsPerChunk, numLeaves, numChunks, treeHeight, tcn);
+                PTHREAD_NUM, (long int)numElements, bytesPerElement, numElsPerChunk, numLeaves, numChunks, treeHeight, tcn);
     }
 
 	numChunks = (size_t)tcn;
@@ -259,7 +259,7 @@ pointer GC_arrayAllocate(GC_state s,
     if (DEBUG_CHUNK_ARRAY)
         fprintf(stderr, "%d]   Initial allocHead "FMTPTR" length: %zd, numChunks = %zd\n",
             PTHREAD_NUM, (uintptr_t)allocHead,
-            getLengthOfList(allocHead),
+            (long int)getLengthOfList(allocHead),
             numChunks);
 
     /* allocate root node */
