+++ runtime/gc/umheap.c
@@ -345,9 +345,9 @@ bool createUMHeap(GC_state s,
 	size_t step = max_chunk_size + sizeof(UM_header); /*account for size of chunktype field*/
 	pointer end = h->start + h->size - step;
 	if (DEBUG) {
-		fprintf(stderr, "%d] sizeof(struct GC_UM_Chunk) = %d\n", PTHREAD_NUM, sizeof(struct GC_UM_Chunk));
-		fprintf(stderr, "    sizeof(struct GC_UM_Array_Chunk) = %d\n", sizeof(struct GC_UM_Array_Chunk));
-		fprintf(stderr, "    final chunk size = %d\n", max_chunk_size);
+		fprintf(stderr, "%d] sizeof(struct GC_UM_Chunk) = %d\n", PTHREAD_NUM, (int)sizeof(struct GC_UM_Chunk));
+		fprintf(stderr, "    sizeof(struct GC_UM_Array_Chunk) = %d\n", (int)sizeof(struct GC_UM_Array_Chunk));
+		fprintf(stderr, "    final chunk size = %d\n", (int)max_chunk_size);
 	}
 	struct timeval t0, t1;
 	gettimeofday(&t0, NULL);
@@ -378,7 +378,7 @@ bool createUMHeap(GC_state s,
 	if (DEBUG or s->controls.messages) {
 		fprintf(stderr, "[GC] Created heap of %d chunks in %lu us step=%d sz(umchunk)=%d sz(umhdr)=%d\n",
 				s->fl_chunks, elapsed,
-				step, sizeof(struct GC_UM_Chunk), sizeof(UM_header));
+				(int)step, (int)sizeof(struct GC_UM_Chunk), (int)sizeof(UM_header));
 		fprintf(stderr,
 				"[GC: Created heap at "FMTPTR
 		" of size %s bytes\n",
